Grant access API conflicts resolved
